+++ src/Git_Updater/Plugin.php
@@ -363,7 +363,15 @@ class Plugin {
 			$transient = new \stdClass();
 		}
 
-		foreach ( (array) $this->config as $plugin ) {
+		/**
+		 * Filter repositories.
+		 *
+		 * @since 10.2.0
+		 * @param array $this->config Array of repository objects.
+		 */
+		$config = apply_filters( 'gu_config_pre_process', $this->config );
+
+		foreach ( (array) $config as $plugin ) {
 				$plugin_requires = $this->get_repo_requirements( $plugin );
 				$response        = [
 					'slug'             => $plugin->slug,